# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MustaFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/versions:
    get:
      operationId: listVersions
      tags: [versions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectVersion"
    post:
      operationId: createVersion
      tags: [versions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectVersionInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectVersion"
  /projects/{id}/versions/{versionId}:
    patch:
      operationId: patchVersion
      tags: [versions]
      summary: Rename a version checkpoint label
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: versionId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchVersionBody"
      responses:
        "200":
          description: Updated version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectVersion"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: getVersion
      tags: [versions]
      summary: Get a single version including full files snapshot
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: versionId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Version detail with files snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectVersionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{id}/versions/{versionId}/rollback:
    post:
      operationId: rollbackVersion
      tags: [versions]
      summary: Restore the project files to a saved version snapshot
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: versionId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollbackInput"
      responses:
        "200":
          description: Rollback applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollbackResult"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ProjectVersion:
      type: object
      required: [id, projectId, label, createdAt, filesCount]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        label: { type: string }
        note: { type: ["string", "null"] }
        changelogEntry: { type: ["string", "null"] }
        filesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        planSnapshot:
          nullable: true
          type: object
          additionalProperties: true
    ProjectVersionInput:
      type: object
      required: [label]
      properties:
        label: { type: string, minLength: 1 }
        note: { type: string }
    PatchVersionBody:
      type: object
      properties:
        label: { type: string, minLength: 1 }
    ProjectVersionDetail:
      type: object
      required: [id, projectId, label, createdAt, filesSnapshot]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        label: { type: string }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        filesSnapshot:
          type: array
          items:
            type: object
            required: [path, content, mimeType]
            properties:
              path: { type: string }
              content: { type: string }
              mimeType: { type: string }
...
    RollbackInput:
      type: object
      properties:
        restoreDb:
          type: boolean
          description: "If true and the version has a linked database snapshot, restore it along with the files."
    RollbackResult:
      type: object
      required: [restoredFiles, versionId, label]
      properties:
        restoredFiles: { type: integer }
        versionId: { type: integer }
        label: { type: string }
        dbSnapshotRestored: { type: boolean }
        dbSnapshotId: { type: ["integer", "null"] }
        dbSnapshotError:
          type: string
          nullable: true
          description: "Set when restoreDb was requested but the database restore could not be completed. Null on success."
...